modify testcases
